% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Functions_forPackage.R
\name{boot.pca.similarity}
\alias{boot.pca.similarity}
\title{Bootstrapping the PCAsimilarity function from the evolqg package}
\usage{
boot.pca.similarity(d, i, matrix.1.pop, matrix.2.pop)
}
\arguments{
\item{d}{dataframe}

\item{i}{iteration number that will be provided by the boot package (you do not enter anything)}

\item{matrix.1.pop}{a character that gives the name of the first population (will be used for filtering)}

\item{matrix.2.pop}{a character that gives the name of the second population (will be used for filtering)}
}
\value{
boot data frame from which you can make further calculations
}
\description{
This function takes a single dataframe that is the combination of the two data frames of interest (example: rbind(pop1.df, pop2,df)).
It is important that the rbind'd single dataframe contains a column that identifies which rows belong to which population, so that the
function can split up the data based on the user entered population name.
This function takes the single dataframe, splits it into two data frames (one per population), bootstraps covariance matrices for each,
and finally calculates the PCAsimilarity score
}
